requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

pandas
kagglehub[pandas-datasets]
snowflake-connector-python
snowflake-sqlalchemy
apache-airflow-providers-snowflake
dbt-core
dbt-snowflake
streamlit
plotly
cryptography
